+++ src/main/java/vsla_admin/userManager/role/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Short> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByRoleName(String username);
 }
